Zmeiniony system strzelania
Dodano:
- system strzelania lepszego prefaba
- system footrpintow
- player respawn

+++ Assets/NetcodeForEntitiesSetup/Scripts/myScripts/Server/Data/PerformanceStats.cs
@@ -3,5 +3,5 @@ using Unity.Entities;
 public struct PerformanceStats : IComponentData
 {
     public float FPS;
-    public int Ping;
+    public float Ping;
 }
